feat(nat): Update stateless NAT context build for port ranges
Now that stateless NAT tables support IP prefixes with associated port
ranges, update the RangeBuilder struct and its implementation to build
the stateless NAT tables accordingly. It's a bit messy, but in the end
we get there.

Note: we convert back the generated PortAddrTranslationValue to a
simpler AddrTranslationValue when possible. This allows simpler (and
more efficient) processing when port ranges are not required. It may be
more correct to convert back _only_ when port ranges are not specified
at all in the config for the relevant VpcExpose. But we'd need a pass
before building the tables, to determine whether any of the prefixes in
the lists of the VpcExpose uses ports (at the moment, we only have this
information for the current prefix being processed, not for the other
original prefixes). We may also want to optimise the RangeBuilder if we
know there is no port range for a given Expose; although it's not
critical, as this code only runs at config build time.

Signed-off-by: Quentin Monnet <qmo@qmon.net>