feat(dataplane): use hostname if no --name arg
The --name cmd line argument is purposedly optional.
In production, the gateway name will match the hostname of the host
where the dataplane runs. This retrieves the hostname from the
OS and sets it as the gateway name if --name is not provided.

With this, the gateway name defaults to the hostname but can always
be overriden with --name.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>