feat(nat): Add Port Address Translation to stateless NAT
After the various complex commits that ensued from the addition of port
ranges to VpcExpose objects, this commit looks pretty straightforward.
If we get a port mapping from the stateless NAT tables lookup, update
the source and/or destination L4 port accordingly for the packet. Port
update is a bit messy because we need to process each protocol
separately; we might want to refactor this in the future.

Signed-off-by: Quentin Monnet <qmo@qmon.net>