build(concurrency-macros): enumerate required features
In general, we try to be explicit about which features we are opting for rather than relying on defaults.
As such, default features (and thus all opt-in features) are disabled at workspace level.

We still do this for compile-time only features.
That said, compile-time feature minimization is much less useful than runtime feature minimization.
More, the features of syn and quote are rather complex and it is just more trouble than it is worth
to opt in on a feature by feature basis.
So we just re-enable defaults at the package level here.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>