feat(nat): Improve target prefix lookup for stateless NAT mapping
The idea behind the stateless NAT lookup is that, for a given IP and
port, we find the matching covering prefix (and associated port range)
in a LPM trie, and get the list of corresponding target ranges from the
LPM value. Then we retrieve the offset of the IP and port within the
prefix (LPM key), and look for the corresponding IP and port within the
target prefixes (LPM value).

To find the IP and port at the right offset, we need to iterate over all
target prefixes in the list and to compare the accumulated size with the
offset. This is not efficient; instead, we can use another data
structure to perform a faster lookup.

To that end, replace the list of prefixes with a DisjointRangesBTreeMap
that associates a target prefix to each portion of the original prefix.

Suggested-by: Manish Vachharajani <manish@githedgehog.com>
Signed-off-by: Quentin Monnet <qmo@qmon.net>