fix(net): Replace ahash with rapidhash for ecmp
Previously we were using ahash to hash packets for ecmp
routing.  However, ahash makes not attempt to be stable
across architectures or even feature flags.

To make sure all gateway instances would hash packets to
the same nexthop when ECMP was in play, there was a test
to ensure that ahash didn't change behavior without anyone
noticing.

With the subsequent commits, a dependency somewhere uses
ahash with std or rng enabled, which changes ahash's
behavior so that each instance of ahash starts with a
random seed breaking the gurantees we want for ECMP.

Because of how rust merges all used features together
when building, the use of std or rng in a different crate
affects the net crate as well. As a result, the ahash
consistency test starts to fail.

This commit replaces ahash with rapidhash which has a
goal of hashing consistently across all platforms and
instances of the hasher.  Accordingly, the test and
artifact to test for ahash consistency is removed as
well.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>