build(nix): introduce nix build system
The goal of this series of commits is to migrate functionality from
dpdk-sys (and give it a much needed refactor in the process).

More specifically, this series of commits is intended to get
the project to the point where you can do a sterile build of DPDK
with the correct set of CFLAGS/LDFLAGS in both the debug and release
profiles.

Forming that build into a sysroot which can actually replace dpdk-sys
will be left as future work.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>