build(nix): introduce stdenv-llvm
This commit introduces a nixpkgs provided stdenv based on the LLVM
toolchain.  We explicitly forswear GCC for this sysroot on the
grounds that rustc is based on LLVM. Given that we want LTO to work
properly, we need to ensure that all of our (modest list of) dataplane
dependencies are built with LLVM.

In a future commit we will add infrastructure to ensure that we are also
always using the same version of LLVM used by our selected version of
rustc.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>