build(nix): add llvm's bintools and lld to stdenv
For reasons which make no sense to me, llvmPackages.stdenv does not
seem to add LLVM's bintools or linker to the nativeBuildInputs by
default.

If you attempt to pass `-fuse-ld=lld` to the compiler without explicitly
adding them to the build time dependency list then you fail to link.

The fix is easy: inject those packages as dependencies.

Note that I deliberately don't use `pkgs.lld` as that version of lld is not
wrapped by nix and will not ensure correct settings for the rpath in generated
elf files.  llvmPackages.lld is the version we need here, as that version of
lld is wrapped.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>