build(nix): disable checks in stdenv
It is neat that nixpkgs attempts to run unit tests for compiled packages,
but in my experience those tests spuriously fail on the regular.
The cause of these failures is almost uniformly due to insufficient
permissions in the build environment.

I will leave the testing of packages to the developers and package maintainers
and simply disable those tests in our overlay.

This is especially important in that we will (in the future) cross compile for
aarch64 and our build machine would be unable to run the tests anyway.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>