build(nix): introduce dpdk package
This commit introduces a build recipe for DPDK optimized for our
use case.

This is largely just transferred over from dpdk-sys.

The primary reason to focus on minimizing DPDK is to make it as easy
as possible to bind DPDK to rust. In particular, enabling all of the
drivers significantly complicates the binding process and
significantly increases build times.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>