build(nix): frame common profile
Now that we have our overlay in a place where it can accept flags
we can start constructing machinery to generate those flags.

We start with the "common" profile.

This profile consists of flags we expect to pass to each and every
build environment on all target platforms. Flags should be added to
this profile conservatively, as they apply globally.  Do not add
performance compromising flags here.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>