build(nix): frame secure profile
The secure profile is a set of flags intended to degrade security
vulnerabilities into denial of service attacks.

This should mostly be used in release builds, but can be applied to
debug or (especially) to (future) fuzz environments as a means of
instructing the fuzzer to hunt bugs with plausible security impacts.

In particular, enabling safe-stack and or cfi (both future goals)
should be done in future fuzz builds if possible.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>