ci: Pass TMPDIR to compile environment (again)
We want to enable running doctests in CI. We're facing the same issue as
we did in the past: the /tmp directory, where rustdoc seems to create
symblinks to the executables for running the tests, is mounted with
"noexec" and the step would fail in CI.

Commit 6e07f4467953 ("test: cleanup pointless logic") removed some logic
that was maybe not so "pointless" after all, since we had passed a
custom TMPDIR path in commit 6d722f17dde2 ("ci: Pass TMPDIR to sterile
environment"), precisely to address this issue. Let's re-apply a similar
change: pass a custom TMPDIR environment variable pointing to some
location where we can actually exec the test binaries.

Fixes: 6e07f4467953 ("test: cleanup pointless logic")
Signed-off-by: Quentin Monnet <qmo@qmon.net>