build(nix): introduce fat-lto-objects flag (for the moment)
This flag allows builds which make use of tools like nm to build when lto is enabled.

We enable it here (for the moment) to allow rdma-core to build on aarch64 / aarch64-musl.

That said, I don't like this flag as it makes the object files and thus the archive files much bigger. It opens the
window for LTO linking mistakes which are much easier avoided if this is off.

Once I have cooked a patch for rdma-core to skip the nm step we can revert this commit.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>