chore: Upgrade to k8s agent CRD 0.32.0
The updated CRD contains group configuration for gateway
failover, though dataplane does not support it.  This
commit just sets those fields to None.  Later commits
can make those configuration options functional.

The CRD also adds a field for a dataplane version which
this commit sets to the value of the VERSION environment
variable set by the justfile during build.

The CRD also adds a last_heartbeat field that indicates
the last time dataplane updated the status.  This is for
consistency with other entities in the Open Network Fabric.
This commit sets that heartbeat to the current time whenever
status is updated.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>