feat(flow-filter): Add new flow-filter stage to enforce peering rules
The new flow-filter crate introduced in this commit contains a new
pipeline stage that addresses multiple purposes:

- It validates that a packet matches an existing peering connection
- It retrieves the destination VPC for the packet
- It improves prefix overlap handling, when compared to the existing
  destination VPC lookup stage

Flow validation
---------------

This flow-filter stage validates that a packet matches an existing
peering connection, as defined in the configuration provided by the
user. All packets that do not have a source IP, port and destination IP,
port corresponding to existing, valid connections between the prefixes
in exposed lists of peerings, get dropped.

This allows us to enforce that traffic matches the peering rule:

- Generically, whether or not the appropriate rules are set
- For stateless NAT, which would just let the packet go un-NATed if
  there was no NAT rule to be found, so far
- For stateful NAT, allowing us to remove the dirty implementation of
  similar checks in the stateful NAT code (in future work).

Destination VPC lookup
----------------------

Since filtering flow involves retrieving the source VPC, source IP/port
and destination IP/port, then looking in a table, we've got everything
we need to reimplement a destination VPC lookup stage. It's trivial to
store the destination VPC in the context table, and to attach it to the
packet's metadata on successful lookups. Doing so, we replicate the
function of the dst_vpcd_lookup stage: the legacy, redundant stage will
be removed in a follow-up commit. This will make one stage fewer in the
pipeline, and will improve accuracy, as described in the next section.

Improved resiliency for "shared IPs"
------------------------------------

When looking up for the destination VPC for a packet, we need a special
treatment when IP/port ranges are shared across peerings. In the legacy
dst_vpcd_lookup stage, we already had the checks in place to make sure
that the lookup stage does not return an answer when there are multiple
possible matching destination VPCs (based on source VPC and destination
address).

However, the restriction in that stage is too strong: when we have
overlapping, but distinct IP ranges shared across peerings, we can tell
the destination VPC for some portion of the ranges.

Consider the following case, for example:

    VPC A        <->        VPC B        <->         VPC C
    1.0.0.0/24     2.0.0.0/24   2.0.0.0/24           1.0.0.0/24

In this case, we just can't tell, for packets coming from VPC B, what
the destination VPC is (unless there's some stateful NAT session in
place, but that's beyond the destination VPC lookup stage's scope).

But if, instead, we have:

    VPC A        <->        VPC B        <->         VPC C
    1.0.0.0/23     2.0.0.0/24   2.0.0.0/24           1.0.0.0/24

Then for packets going to 1.0.1.0/24, we know the destination is VPC A.
Before this commit, the destination VPC lookup would not account for the
unambiguous portion of prefixes.

Similarly, in a configuration such as this one:

    VPC A        <->        VPC B        <->         VPC C
    1.0.0.0/24     5.0.0.0/24   2.0.0.0/24           1.0.0.0/24

Then we can find the relevant destination VPC for packets coming from B,
because even if the destination prefixes are the same, we can look at
the source address too, and find which (non-overlapping, in our example)
prefix it belongs to. Before this commit, the destination VPC lookup
would never account for the source address.

This commit does not change the legacy stage, but the flow-filter stage
is able to identify the correct destination VPC in both situations.

Unit tests for the new pipeline stage will come in a follow-up commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>