feat(flow-filter): Add new flow-filter stage to enforce peering rules
The new flow-filter crate introduced in this commit contains a new
pipeline stage for validating that a packet matches an existing peering
connection, as defined in the configuration provided by the user. All
packets that do not have a source IP, port and destination IP, port
corresponding to existing, valid connections between the prefixes in
exposed lists of peerings, get dropped.

This allows us to enforce that traffic matches the peering rule:

- Generically, whether or not the appropriate rules are set
- For stateless NAT, which would just let the packet go un-NATed if
  there was no NAT rule to be found, so far
- For stateful NAT, allowing us to remove the dirty implementation of
  similar checks in the stateful NAT code (in a future commit).

Unit tests also come in a follow-up commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>