refactor(lpm): Use enum for PrefixWithOptionalPorts
Using an Option<PortRange> to make the port range optional may lead to
confusion, in particular when we attempt to get the .size() of the
object. It's cleaner to use an enum instead, which can take either a
Prefix or a PrefixWithPorts (reusing the latter's methods to implement
equivalent methods for PrefixWithOptionalPorts) - even though it's a bit
more verbose overall.

Suggested-by: Manish Vachharajani <manish@githedgehog.com>
Signed-off-by: Quentin Monnet <qmo@qmon.net>