feat(mgmt): add ConfigClient
The ConfigClient is the entity used to interact with the
config processor. This abstraction will allow us to simplify
configuration entry points (grpc/k8s or any other), hiding the
internals of the communication that are now unnecessarily
exposed.

This is the first commit in a series to simplify the mgmt crate.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>