feat(flow-filter): Split overlapping prefixes for shared IPs support
When looking up for the destination VPC for a packet, we need a special
treatment when IP/port ranges are shared across peerings. We already
have the checks in place to make sure that the lookup stage does not
return an answer when there are multiple possible matching destination
VPCs (based on source VPC and destination address).

However, the restriction is too strong: when we have overlapping, but
distinct IP ranges shared across peerings, we can tell the destination
VPC for some portion of the ranges.

Consider the following case, for example:

    VPC A        <->        VPC B        <->         VPC C
    1.0.0.0/24     2.0.0.0/24   2.0.0.0/24           1.0.0.0/24

In this case, we just can't tell, for packets coming from VPC B, what
the destination VPC is (unless there's some stateful NAT session in
place, but that's beyond the destination VPC lookup stage's scope).

But if, instead, we have:

    VPC A        <->        VPC B        <->         VPC C
    1.0.0.0/23     2.0.0.0/24   2.0.0.0/24           1.0.0.0/24

Then for packets going to 1.0.1.0/24, we know the destination is VPC A.

Right now the destination VPC lookup does not account for the
unambiguous portion of prefixes.

This commit does not change that, but it improves the flow-filter stage
instead, which is used to determine whether packets belong to allowed
connections (as defined by the user's peering configuration). This is
not necessary to check that the packet is allowed (if multiple valid
connections are present, the packet is definitely allowed), but due to
the context present in the flow-filter table, we consider replacing the
destination VPC lookup stage by the flow-filter stage only. This would
make one less stage in the pipeline, and would address the shared IPs
issues described above.

Signed-off-by: Quentin Monnet <qmo@qmon.net>