feat(flow-filter): Set destination VPC for packet from flow-filter
In preparation for replacing the destination VPC lookup stage, make the
flow-filter set the destination VPC in the packet's metadata.

The expectation is that the flow-filter stage should always find the
same destination VPC as the dst_vpcd_lookup stage, with additional
precision in the case of overlapping but distinct prefixes shared across
peerings. The dst_vpcd_lookup stage is not remove yet.

Signed-off-by: Quentin Monnet <qmo@qmon.net>