feat(k8s-less): add k8s-less crate
Adds a new crate to support kubeless mode (k8s-less), where
dataplane learns its configuration by monitoring changes in files
inside a directory (using inotify) that contain the gateway spec
in yaml or json.

Files that change in (or get added to) the watched directory get
read and their contents (yaml or json) deserialized into a crd
spec. From that spec, a GatewayAgent object (representing the main
CRD) is built and passed to a user-provided callback.

This crate attempts to mimic k8s-intf and deals only with:
  - the watching of files in a directory using inotify
  - the deserialization of those files into the GatewayAgent type.

It does not attempt to populate/build any of the configuration
types in crate dataplane-config.

As this mode of operation may be used only for testing or
development, we implement it outside of the k8s-intf crate.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>